move and refactor compat
